backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.10.4
pydantic-settings==2.7.0
convex==0.6.0

# PDF parsing
pymupdf==1.25.1

# Embeddings & Vector Store
numpy==2.2.3
google-generativeai==0.8.3

# Reference validation
httpx==0.28.1
rapidfuzz==3.11.0

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
asyncio-throttle==1.0.2
tenacity==9.0.0
slowapi==0.1.9
